+++ src/app/layout.tsx
@@ -1,8 +1,11 @@
 import './globals.css'
 
 export const metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: {
+    default: 'imgs',
+    template: '%s - imgs',
+  },
+  description: 'imgs is Image Viewer on Secure Cloud Storage used end-to-end encryption.',
 }
 
 export default function RootLayout({
